requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
lxml==5.2.2
pandas==2.2.2  # optional; remove if you don’t want it
